[TOPI] Expose the interface of `topi.collapse_sum` (#102)
TOPI has an implementation of `collapse_sum` internally
(`tvm/topi/reduction.h`) but it is not exposed to FFI and can not be
called in Python side. This patch exposes it and adds some related
tests.
Besides, now legalizer can legalize `collapse_sum_like/to`! But due to
the TOPI implementation, it can't handle the symbolic case.