[FIX][AD] Fix nested msg utils introduced in AD Pass (#109)
This PR is a small fix patch for
https://github.com/mlc-ai/relax/pull/103, containing two small
modifications:
- In AD PR we introduce `NestedMsgToExpr`, where the signature of
`fmapleaf` is `Expr fmapleaf(T)`. And for null nested msg it will
directly throws error. Now we change it to `Expr fmapleaf(Optional<T>)`
and pass `NullOpt` to `fmapleaf`, which enables user to decide whether
to throw an error or return some default value.
- In the test of nested msg we forget to change the signature of
`fmapleaf` (originally it is `Expr fmapleaf(NestedMsg<T>)`). It still
passed due to the type casting between `NestedMsg<T>` and `T` but it
needs to be fixed.