POC of LiftTransformParams (#122)
This is a prototype of `LiftTransformParams`. It allows to compile the
end-to-end model without weights provided. The idea is annotate the
input parameters that are weights, and identify and lift the
transformations to weights, and compile it to a separate function
`transform_params` that can be executed in runtime. Users can run
`transform_params` with weights to get the weights for the optimized
model as a prep step before the deployment. In this way, we perform the
same optimizations and defer the weight transformations to the user
side, while the overhead of the deferred weight transformation can be
ignored as it only need to be run once.

A demo notebook is available
[here](https://github.com/vinx13/relax/blob/fda857553557c34f97a4f7193a529da607a3421c/tests/python/relax/demo_lift_transform_params.ipynb)

This pass is not integrated yet with the default `vm.build` as we are
going to iterate it.