[Transform] Refactor LegalizeOps (#128)
@MasterJH5574 brings the very first version of legalizer for relax
high-level operators in this PR:
https://github.com/mlc-ai/relax/pull/96. This PR refactors the
`LegalizeOps` pass according to the suggestions mentioned in the
comments of the PR. In details, it includes:
- Change the default legalization map to the attribute of operators. By
this way we can make better use of op attribute and move the main body
of the Pass to C++, making `LegalizeOps` a more formal Pass and bringing
opportunities that we can register legalization without Python in the
future.
- Use decorator to register legalization instead of writing it in the
Pass. It is more elegant and brings extensibility. Also by this way we
can have a clear code structure since now we separate the implementation
in different files. As we will have more and more legalization for
operators, putting them in a single file is not promising.
- Preserve the customize map part. It looks good for now.