[Op][Training] Support Batch Norm Training Mode (#142)
### Background

Operator `batch_norm` behaves differently in eval mode and train mode.
In training, it will maintain two internal states `moving_mean` and
`moving_var` while these updates don't appear in eval mode.

In
[torch](https://pytorch.org/docs/stable/generated/torch.nn.functional.batch_norm.html),
we have a flag `training: bool` to indicate this and a param `momentum`
for updating. So it is necessary for us to introduce them to achieve
training with bn.

What's more, consider we have a forward function with `batch_norm`
operator in eval mode inside. If we have an AD pass, we may want to
differentiate it and get the function which can be used to train. But
the problem is that `batch_norm` is still in eval mode. So a pass is
needed to convert all operators inside a function / mod from eval mode
to training mode, by setting their flag `training` as `true`.

### Work

This PR adds `training: bool` and `momentum: float` arguments for batch
norm in Relax and TOPI. And it contains a Pass named `ToTrainMode` which
converts all operators to training mode.