[Pass] Simplify Batch Norm in Training Mode (#162)
**Background**

Prior to this PR, we have the pass `SimplifyNormInference` which helps
us unpack the batch norm computation in index 0. But since the gradient
of BN is complex and a good solution is to simplify it to smaller
operators, it is necessary that we should introduce simplify for
training, not just for inference.

**Work**

This PR changes the pass `SimplifyNormInference` to `SimplifyNorm`, and
add an argument `mode: Literal["eval", "training"]` to indicate the
simplification mode.