[Training] Bind TE Gradient Func to TIR Name (#195)
This PR provides an interface to let user bind a manually written TE
gradient function to a TIR function name. This binding information will
be stored as an attribute of the module (with attr key
`te_grad_bind_handler`).
When the gradient pass meet a `call_tir`, it will try to search whether
there is a TE gradient function bound to this name. If so, it will call
the handler (i.e. A wrapped function to execute
`bb.emit_te(te_grad_func, ...)`) and get the emitted var. Then the
partials will be set to `var[0], var[1], ...` and we can do the adjoints
propagation.