[Transform] Enhance FuncCopier (#225)
Now FuncCopier will provide a var_map member as the relax var mapping
from the old function and the new function.

In FuncCopier we will call SymbolicVarRenewMutator, which also renews
relax Vars. However, previously we did not consider this var mapping
into account, leading that we will provide wrong var_map in FuncCopier.

That leads to bugs in the Gradient pass, whichs uses this member.

This PR will merge the var mappings in FuncCopier and in
SymbolicVarRenewMutator.